wip context refactor
