+++ Tests/ToucanSDKTests/Utilities/MinifyHTMLTests.swift
@@ -15,8 +15,6 @@ final class MinifyHTMLTests: XCTestCase {
         let html =
             "<html>   <body>   <h1>Hello, world!</h1>   </body>   </html>"
         let minifiedHTML = html.minifyHTML()
-        //        print(minifiedHTML)
-
         XCTAssert(!minifiedHTML.isEmpty)
     }
